feat(db): add 3 perf indexes for hot read paths
Three queries fell to full-scan as chain volume grew. None matter
today (52K mainnet txs / 1.2M testnet) but they'd start to hurt
around 1M+ txs. Add B-tree indexes now while the migration is cheap.

  - addresses_contract_recent_idx (is_contract, first_seen_block)
    serves /contracts/recent: WHERE is_contract = true ORDER BY
    first_seen_block DESC
  - txs_value_desc_idx (value)
    serves /whale/tx: ORDER BY value DESC LIMIT N
  - cb_trades_srx_amount_desc_idx (srx_amount)
    serves /coinblast/whales: WHERE type != 'graduated' ORDER BY
    srx_amount DESC LIMIT N

Also drop the redundant ::numeric casts in /whale/tx — value is
already numeric(78,0); the cast wrapped the column expression and
made the planner refuse the index. Without the cast, plan flips
from Parallel Seq Scan (cost ~58k on testnet) to Index Scan
Backward (cost ~1).

Migration is non-CONCURRENTLY (drizzle migrator wraps each .sql in
a transaction; CONCURRENTLY can't run inside one). Tables are
small enough that the brief ACCESS EXCLUSIVE lock is acceptable.
Operators on a giant DB later can run these by hand with
CONCURRENTLY outside the migrator.

Verified on testnet — all three indexes created, /whale/tx plan
verified flipped to Index Scan Backward.

+++ apps/api/src/routes/native.ts
@@ -321,6 +321,11 @@ export function registerNativeRoutes(
         try { parseBigIntOrThrow(threshold, "threshold"); }
         catch (e) { return reply.code(400).send({ error: (e as Error).message }); }
       }
+      // Drop the redundant `::numeric` casts on `value` — column is
+      // already numeric(78,0). Wrapping the column in a cast made the
+      // planner treat ORDER BY as an expression and refuse the
+      // txs_value_desc_idx B-tree (parallel seq scan, ~58k cost). With
+      // the casts gone the planner uses Index Scan Backward (~1 cost).
       const rows = threshold
         ? await ctx.db.execute<{
             hash: string;
@@ -335,8 +340,8 @@ export function registerNativeRoutes(
                      t.block_height::text, b.timestamp::text
               FROM ${transactions} t
               JOIN ${blocks} b ON b.height = t.block_height
-              WHERE t.value::numeric >= ${threshold}::numeric
-              ORDER BY t.value::numeric DESC, t.block_height DESC
+              WHERE t.value >= ${threshold}::numeric
+              ORDER BY t.value DESC, t.block_height DESC
               LIMIT ${limit}
             `
           )
@@ -353,7 +358,7 @@ export function registerNativeRoutes(
                      t.block_height::text, b.timestamp::text
               FROM ${transactions} t
               JOIN ${blocks} b ON b.height = t.block_height
-              ORDER BY t.value::numeric DESC, t.block_height DESC
+              ORDER BY t.value DESC, t.block_height DESC
               LIMIT ${limit}
             `
           );
